Fix NSScanner overflowing NSDecimal exponent
And consume the correct number of characters even if the exponent looks too large.

This was causing window max size to be read wrong, and that caused all the subviews
to resize to some bogus frames. That could result in them visually moving on the
screen and/or stopping to receive mouse events.