Fix crashing when deallocating a run loop before its delayed performers are finalized
For some reason, CF prefers to CFRelease() the NSRunLoop -- which, by the way, is not a correct way to release
a non-toll-free-bridged object such as NSRunLoop, but it happens to work because of the way toll-free bridging
is implemented -- before finalizing all the timers, which means our NSDelayedPerformer's stay alive a bit longer
then the loop they belong to. This is fine, because they get released immediately after that, but we should keep
the two timers arrays alive long enough.