Fix calling convention issues in NSGeometry
Regular and vararg function prototypes/pointers are *not* interchangable;
floats get passed differently. objc_msgSend() and IMP-s are (normally)
regular non-vararg functions, despite their compiler-percieved types
indicating otherwise; so it's important to cast them to the appropriate
function pointer types before invoking them. The reverse is also true:
+[NSString stringWithFormat:] is a vararg method, but without NSString.h,
the compiler assumes it to be a regular method and uses an incorrect ABI
for passing floats.

Fix the latter by including NSString.h and the former by porting back the
GNUstep caching code that casts the method implementation types properly.
Unlike on GNUstep, our scanner is actually an NSConcreteScanner, so account
for that as needed.