Hide use of C11 atomics in NSConnection and NSDistantObject
Including `stdatomic.h` conflicts with C++'s `atomic`. "Why is this a problem?", you ask? Well, because this little thing called Objective-C++ exists, and Foundation needs to play nice with C++ for that reason.

Note that the current fix is *technically* UB (though Clang is supposed to do the right thing when working with integral types, as we are), so if we ever run into issues with it (doubtful, but still), you might want to look into Clang's `__atomic_*` builtins.