Fix non-shared NSUserDefaults
The old code was crashing when individual NSUserDefaults instances were created without ever accessing the shared one since the shared one initialized the shared synchronization timer.

Having this timer and the associated queues be shared is wrong anyways, particularly for the timer, as it was receiving the shared instance as the `self` parameter and would thus ignore any custom `suiteName` for individual instances of NSUserDefaults.