Many fixes for NSBundle
First of all, implement allFrameworks.

Second of all, according to Apple's documentation, bundles have to be unique. This means we need to keep track of them in some global table keyed by identifiers and/or paths. The init functions are actually guaranteed to return an existing bundle and deallocate self if another bundle object exists for the given identifier or path.