Theoretical service listener support
Like with NSProgress support, this is theoretical because it can't be properly tested at the moment. Once our `xpc_main` is in working order (i.e. once it can launch bundled services), then this can be tested.

Also, NSUUID *does* support secure coding, so make sure it reports that (so it can be sent over NSXPC).