Implement NSTask
Note: there might be complications with the way that waiting is implemented in waitUntilExit. Namely, if the CFRunLoop we're waiting on enters another nested execution of itself, the wake-up will only exit that inner loop. During testing, this hasn't been an issue, but keep in mind that it might cause problems.