Properly fix `readDataOfLength:` to respect limit
My last commit modified `bufSize` to try to respect the length limit.
The problem with this is that `bufSize` is automatically adjusted when
the buffer is filled up. So instead, modify `READ_SIZE` to respect the
limit. This also requires a change to `maxReadSize`, otherwise each
iteration of the loop uses the minimum of the unmodified `READ_SIZE`
and the modified bufSize (we need to take `totalReadSize` into
account on both).