Fix NSUnarchiver readType failing to decode a shared string
The call to decodeString doesn't succeed in the case when charValue is
NewLabel, as it then calls decodeSharedString, which then tries to check
the next character for NullLabel/NewLabel/Default, fails as it is now on
the first character of the string.

I'm not for sure if decodeString is supposed to be called internally in
NSUarchiver. Inside NSUarchiver it is calling decodeSharedString every
place else. The difference between decodeString and decodeSharedString
is very minor.

I opted to just call decodeShareString in readType and it fixed the
described issue above.